+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Reduced the size of the icons and clear button for compressed `EuiFormControlLayout` ([#4374](https://github.com/elastic/eui/pull/4374))
 - Added ability for text input updates in `EuiDatePicker` ([#4243](https://github.com/elastic/eui/pull/4243))
-- Added a `size` prop to `EuiContextMenu` and `EuiContextMenuPanel` ([#4409](https://github.com/elastic/eui/pull/4409))
+- Added a `size` prop to `EuiContextMenu` and added a smaller size ([#4409](https://github.com/elastic/eui/pull/4409))
 
 ## [`31.0.0`](https://github.com/elastic/eui/tree/v31.0.0)